separated Search into its own component so it could be used on Profile page
